Add abstract class Node
